refactor(app): legacy Controller 삭제
