+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -14,7 +14,6 @@ public class AppWebApplicationInitializer implements WebApplicationInitializer {
     @Override
     public void onStartup(final ServletContext servletContext) {
         final var dispatcherServlet = new DispatcherServlet();
-        dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
         dispatcherServlet.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse"));
 
         final var dispatcher = servletContext.addServlet("dispatcher", dispatcherServlet);
